requirements.txt
ansicon==1.89.0
appdirs==1.4.4
attrs==23.2.0
blessed==1.20.0
click==8.1.7
colorama==0.4.6
deepface==0.0.82
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
flyingcircus==0.1.4.1
future==0.18.3
haar-cascade-nms==1.1.1
jinxed==1.2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyproject-toml==0.0.10
pyserial==3.5
python-dateutil==2.8.2
pytk==0.0.2.1
pytz==2023.3.post1
referencing==0.33.0
rpds-py==0.18.0
setuptools==69.0.3
setuptools-scm==8.0.4
six==1.16.0
tk==0.1.0
tkintertable==1.3.3
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
wheel==0.42.0
